requirements.txt
# ── Core Data ──────────────────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4

# ── Machine Learning ──────────────────────────────────────────────────────────
scikit-learn==1.4.2
xgboost==2.0.3
joblib==1.4.0
scipy==1.13.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1

# ── API ───────────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
openpyxl==3.1.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
slowapi==0.1.9

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit==1.33.0
plotly==5.21.0

# ── LLM ───────────────────────────────────────────────────────────────────────
openai==1.25.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
faker==24.4.0
apscheduler==3.10.4
requests==2.31.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.0
httpx==0.27.0
